+++ lib/models/weather.dart
@@ -31,17 +31,17 @@ class Weather with ChangeNotifier {
 
   factory Weather.fromJson(Map<String, dynamic> json) {
     return Weather(
-      temp: json['main']['temp'],
-      tempMax: json['main']['temp_max'],
-      tempMin: json['main']['temp_min'],
+      temp: (json['main']['temp']).toDouble(),
+      tempMax: (json['main']['temp_max']).toDouble(),
+      tempMin: (json['main']['temp_min']).toDouble(),
       lat: json['coord']['lat'],
       long: json['coord']['lon'],
-      feelsLike: json['main']['feels_like'],
+      feelsLike: (json['main']['feels_like']).toDouble(),
       pressure: json['main']['pressure'],
       description: json['weather'][0]['description'],
       currently: json['weather'][0]['main'],
       humidity: json['main']['humidity'],
-      windSpeed: json['wind']['speed'],
+      windSpeed: (json['wind']['speed']).toDouble(),
       cityName: json['name'],
     );
   }
